crimson/os/seastore: split out transaction_interruptor.h
