+++ src/crimson/os/seastore/CMakeLists.txt
@@ -4,6 +4,7 @@ set(crimson_seastore_srcs
   segment_manager.cc
   segment_manager/ephemeral.cc
   segment_manager/block.cc
+  transaction_interruptor.cc
   transaction_manager.cc
   transaction.cc
   cache.cc
